# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NWMS API",
        "description": "This API uses Bearer Token authentication.\n *     To authorize, use the 'Authorize' button above and enter a token in the format:\n *     Bearer {your_token}",
        "version": "1.0.0"
    },
    "paths": {
        "/api/acceptances": {
            "get": {
                "tags": [
                    "Acceptances"
                ],
                "summary": "Get list of acceptances",
                "operationId": "8c248943c80dc486e20bf5c48e7429fc",
                "responses": {
                    "200": {
                        "description": "List of acceptances"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Acceptances"
                ],
                "summary": "Create a new acceptance",
                "operationId": "82e6ba0810aff8ac151e5e778a5dc788",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Acceptance"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Acceptance created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/acceptances/{id}": {
            "get": {
                "tags": [
                    "Acceptances"
                ],
                "summary": "Get acceptance by ID",
                "operationId": "0ec69dce95755d01aa385ca16c1d741a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Acceptance found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Acceptances"
                ],
                "summary": "Update acceptance",
                "operationId": "5748612e57bf9b0d00b215ce8196d5cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/schemas/Acceptance"
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Acceptances"
                ],
                "summary": "Delete acceptance",
                "operationId": "106ec3316623eb4e10b532fd72c1959e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/acceptances/{id}/offers": {
            "post": {
                "tags": [
                    "Acceptances"
                ],
                "summary": "Add offers to an acceptance",
                "operationId": "1471c7e08213c19ff93448a726bef609",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/AcceptanceOfferInput"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Offers added successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/acceptances/{acc_id}/offers/{ao_id}": {
            "put": {
                "tags": [
                    "Acceptances"
                ],
                "summary": "Update an offer in an acceptance",
                "operationId": "b74c419694cda7653245600c619e8ab7",
                "parameters": [
                    {
                        "name": "acc_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ao_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AcceptanceOfferInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Offer updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Acceptances"
                ],
                "summary": "Delete an offer from an acceptance",
                "operationId": "554710da0159ad8ac605d572e5ca24a5",
                "parameters": [
                    {
                        "name": "acc_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ao_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offer deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "operationId": "78ce1bd22726e99d8f30c6c8fa329168",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create new products",
                "operationId": "1e5b76a5975af81b35765edfe66f34ec",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Product"
                },
                "responses": {
                    "201": {
                        "description": "Products created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by ID",
                "operationId": "f4e69191da6baa0f13eb46ff6dd761ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "6f191ad01c85b6853580d01bd9c285cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/Product"
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "operationId": "b883dde0dd648a32e73f7b73699a2a85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get list of orders",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "List of orders"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "61f0ce377b9ca021577dd300752ec6b8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order by ID",
                "operationId": "88d27387b1e409e075c18609d4a0c300",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order found"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an order",
                "operationId": "38a1b85e8294795a1b0dc438a36e319f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order",
                "operationId": "bab875a1ed962fb33565ab9b48c0d494",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{id}/offers": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Add offers to an order",
                "operationId": "54cba00977a8c0431504e25c15fdecf5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/OrderOffer"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Offers added to order"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{order_id}/offers/{offer_id}": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an offer in order",
                "operationId": "35845a5f7aaba22410c40bda6a691d23",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderOffer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Offer updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an offer from an order",
                "operationId": "c4e8eee91c87c6638966e050118c4625",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offer removed from order"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Order": {
                "required": [
                    "o_status_id",
                    "o_type_id",
                    "o_user_id",
                    "o_shop_id",
                    "o_wh_id",
                    "o_date",
                    "o_count",
                    "o_sum"
                ],
                "properties": {
                    "o_status_id": {
                        "type": "integer"
                    },
                    "o_type_id": {
                        "type": "integer"
                    },
                    "o_user_id": {
                        "type": "integer"
                    },
                    "o_ext_id": {
                        "type": "string",
                        "maxLength": 30
                    },
                    "o_shop_id": {
                        "type": "integer"
                    },
                    "o_wh_id": {
                        "type": "integer"
                    },
                    "o_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "o_date_send": {
                        "type": "string",
                        "format": "date"
                    },
                    "o_source_id": {
                        "type": "integer"
                    },
                    "o_count": {
                        "type": "number"
                    },
                    "o_sum": {
                        "type": "number"
                    },
                    "o_operation_user_id": {
                        "type": "integer"
                    },
                    "o_order_place": {
                        "type": "integer"
                    },
                    "o_current_pallet": {
                        "type": "integer"
                    },
                    "o_current_box": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "AcceptanceOffer": {
                "required": [
                    "ao_acceptance_id",
                    "ao_offer_id",
                    "ao_expected"
                ],
                "properties": {
                    "ao_id": {
                        "description": "Primary key",
                        "type": "integer"
                    },
                    "ao_acceptance_id": {
                        "description": "Acceptance ID",
                        "type": "integer"
                    },
                    "ao_offer_id": {
                        "description": "Offer ID",
                        "type": "integer"
                    },
                    "ao_expected": {
                        "description": "Expected quantity",
                        "type": "number",
                        "format": "float"
                    },
                    "ao_accepted": {
                        "description": "Accepted quantity",
                        "type": "number",
                        "format": "float"
                    },
                    "ao_placed": {
                        "description": "Placed quantity",
                        "type": "number",
                        "format": "float"
                    },
                    "ao_price": {
                        "description": "Purchase price",
                        "type": "number",
                        "format": "float"
                    },
                    "ao_batch": {
                        "description": "Batch number",
                        "type": "string"
                    },
                    "ao_expiration_date": {
                        "description": "Expiration date",
                        "type": "string",
                        "format": "date"
                    },
                    "ao_barcode": {
                        "description": "Barcode",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AcceptanceOfferInput": {
                "required": [
                    "ao_offer_id",
                    "ao_expected"
                ],
                "properties": {
                    "ao_offer_id": {
                        "description": "ID of the offer",
                        "type": "integer"
                    },
                    "ao_expected": {
                        "description": "Expected quantity",
                        "type": "number",
                        "format": "float"
                    },
                    "ao_price": {
                        "description": "Purchase price",
                        "type": "number",
                        "format": "float"
                    },
                    "ao_batch": {
                        "description": "Batch number",
                        "type": "string",
                        "maxLength": 30
                    },
                    "ao_expiration_date": {
                        "description": "Expiration date",
                        "type": "string",
                        "format": "date"
                    },
                    "ao_barcode": {
                        "description": "Barcode",
                        "type": "string",
                        "maxLength": 64
                    }
                },
                "type": "object"
            },
            "Acceptance": {
                "required": [
                    "acc_wh_id",
                    "acc_shop_id",
                    "acc_user_id",
                    "acc_type",
                    "acc_status"
                ],
                "properties": {
                    "acc_id": {
                        "description": "Primary key",
                        "type": "integer"
                    },
                    "acc_domain_id": {
                        "description": "Domain ID",
                        "type": "integer"
                    },
                    "acc_status": {
                        "description": "Status ID",
                        "type": "integer"
                    },
                    "acc_ext_id": {
                        "description": "External ID",
                        "type": "integer"
                    },
                    "acc_user_id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "acc_wh_id": {
                        "description": "Warehouse ID",
                        "type": "integer"
                    },
                    "acc_shop_id": {
                        "description": "Shop ID",
                        "type": "integer"
                    },
                    "acc_type": {
                        "description": "Acceptance type",
                        "type": "integer"
                    },
                    "acc_comment": {
                        "description": "Comment",
                        "type": "string"
                    },
                    "acc_count_expected": {
                        "description": "Total expected quantity",
                        "type": "number",
                        "format": "float"
                    },
                    "acc_count_accepted": {
                        "description": "Total accepted quantity",
                        "type": "number",
                        "format": "float"
                    },
                    "acc_count_placed": {
                        "description": "Total placed quantity",
                        "type": "number",
                        "format": "float"
                    },
                    "acc_date": {
                        "description": "Acceptance creation date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "offers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AcceptanceOffer"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
            "Product": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}